+++ cms/src/features/signup/SignUp.tsx
@@ -1,0 +1,14 @@
+import React from "react";
+
+import SignUpForm from "./components/SignUpForm";
+
+const SignUp: React.FC = () => {
+  return (
+    <section>
+      <h2>Sign in</h2>
+      <SignUpForm />
+    </section>
+  );
+};
+
+export default SignUp;