+++ server/src/controllers/auth.controller.ts
@@ -2,12 +2,17 @@ import { RequestHandler } from "express";
 import asyncHandler from "express-async-handler";
 import passport from "passport";
 import jwt from "jsonwebtoken";
+import bcrypt from "bcryptjs";
 
 import validateLogin from "../validators/login.validator";
-import { User } from "../interfaces/user";
+import { validateSignUp } from "../validators/validators";
+import { User, CreateUser } from "../interfaces/user";
+import { matchedData } from "express-validator";
+import { createUser } from "../services/user";
 
 interface authController {
   login: RequestHandler[];
+  signup: RequestHandler[];
 }
 
 const authController: authController = {
@@ -39,6 +44,28 @@ const authController: authController = {
           });
         }
       )(req, res, next);
+    }),
+  ],
+  signup: [
+    validateSignUp(),
+    asyncHandler(async (req, res, next) => {
+      const { password, ...rest } = matchedData<CreateUser>(req, {
+        onlyValidData: true,
+      });
+
+      // TODO
+      // When would hashedPassword be undefined?
+      bcrypt.hash(password, 10, async (err, hashedPassword) => {
+        if (err || !hashedPassword) return next(err);
+
+        const user = await createUser({ password: hashedPassword, ...rest });
+
+        req.login(user, { session: false }, (err) => {
+          const expiresIn = 30; // seconds
+          const token = jwt.sign({ user }, "secretKey", { expiresIn });
+          return res.json({ token });
+        });
+      });
     }),
   ],
 };