+++ server/src/interfaces/user.ts
@@ -21,6 +21,7 @@ type Role = "USER" | "AUTHOR" | "ADMIN";
 
 export interface CreateUser {
   username: string;
+  email: string;
   password: string;
 }