+++ server/src/routes/auth.route.ts
@@ -2,7 +2,7 @@ import { Router } from "express";
 import authController from "../controllers/auth.controller";
 
 const authRoutes = () => {
-  const { login } = authController;
+  const { login, signup } = authController;
   const authRouter = Router();
 
   // GET requests
@@ -10,6 +10,7 @@ const authRoutes = () => {
   // POST requests
   // curl -w "\n" -X POST http://localhost:3001/auth
   authRouter.post("/login", login);
+  authRouter.post("/signup", signup);
 
   // curl -w "\n" -X POST http://localhost:3001/pizza/:id
   // authRouter.post("/logout", getPizza);